feat: add deterministic GM emission simulation tool and GOD panel integration (state accessors, UI hooks, and JSON copy)
