+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/gm_emission_sim.html
@@ -12,6 +12,7 @@
       .status { font-weight: 800; padding: 8px 10px; border-radius: 8px; display: inline-block; }
       .pass { background: rgba(0, 180, 0, 0.15); border: 1px solid rgba(0, 180, 0, 0.35); }
       .fail { background: rgba(220, 0, 0, 0.15); border: 1px solid rgba(220, 0, 0, 0.35); }
+      .warn { background: rgba(245, 158, 11, 0.15); border: 1px solid rgba(245, 158, 11, 0.35); }
       pre { padding: 12px; border-radius: 10px; border: 1px solid rgba(127,127,127,0.35); overflow: auto; }
       ul { margin: 10px 0 14px; padding-left: 20px; }
       code { font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, "Liberation Mono", "Courier New", monospace; }
