+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/index.html
@@ -288,6 +288,16 @@
       <div id="god-analysis-output" class="help" style="font-size:12px; color:#cbd5e1; margin:4px 0 4px 0;"></div>
     </div>
 
+    <!-- GM Emission Sim -->
+    <div class="god-section">
+      <div class="inv-subtitle">GM Emission Sim</div>
+      <div class="god-row" style="margin-bottom:6px;">
+        <button id="god-run-gm-emission-sim-btn" class="btn" title="Run deterministic GM emission simulation scenarios">Run GM Emission Sim</button>
+        <button id="god-copy-gm-emission-sim-btn" class="btn" title="Copy the latest GM emission sim report JSON to clipboard">Copy JSON</button>
+      </div>
+      <textarea id="god-gm-emission-sim-output" readonly style="width:100%; min-height:180px; padding:6px 8px; background:#0b1120; color:#e5e7eb; border:1px solid #334155; border-radius:6px; font-family:'JetBrains Mono', monospace; font-size:11px;"></textarea>
+    </div>
+
     <!-- Smoke Tests -->
     <div class="god-section">
       <div class="inv-subtitle">Smoke Tests</div>
